+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -7,6 +7,8 @@
 ### Breaking Changes
 
 ### Bugs Fixed
+- Fix io counters import issue in performance counters
+  ([#45286](https://github.com/Azure/azure-sdk-for-python/pull/45286))
 - Fix fallback logic when connection string is provided only via environment variable
   ([#45252](https://github.com/Azure/azure-sdk-for-python/pull/45252))